% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{featureCount_repl1}
\alias{featureCount_repl1}
\title{A matrix of feature counts for replicated samples - replicate 1}
\format{
An object of class \code{data.frame} with 30865 rows and 136 columns.
}
\usage{
featureCount_repl1
}
\description{
featureCount_repl1 - a matrix of feature counts for replicated samples - replicate 1
rows are feature; columns are samples.
}
\keyword{datasets}
