% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{featureCount_repl2}
\alias{featureCount_repl2}
\title{A matrix of feature counts for replicated samples - replicate 2}
\format{
An object of class \code{data.frame} with 30865 rows and 136 columns.

An object of class \code{data.frame} with 30865 rows and 136 columns.
}
\usage{
featureCount_repl2

featureCount_repl2
}
\description{
featureCount_repl2 - a matrix of feature counts for replicated samples - replicate 1
rows are feature; columns are samples.

featureCount_repl2 - a matrix of feature counts for replicated samples - replicate 1
rows are feature; columns are samples.
}
\keyword{datasets}
