% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{featureCount_sub1}
\alias{featureCount_sub1}
\title{A matrix of feature counts for a random subset  of samples balanced according to pam50 category}
\format{
An object of class \code{data.frame} with 30865 rows and 627 columns.

An object of class \code{data.frame} with 30865 rows and 627 columns.

An object of class \code{data.frame} with 30865 rows and 627 columns.

An object of class \code{data.frame} with 30865 rows and 627 columns.

An object of class \code{data.frame} with 30865 rows and 627 columns.

An object of class \code{data.frame} with 30865 rows and 627 columns.

An object of class \code{data.frame} with 30865 rows and 627 columns.

An object of class \code{data.frame} with 30865 rows and 627 columns.

An object of class \code{data.frame} with 30865 rows and 627 columns.

An object of class \code{data.frame} with 30865 rows and 627 columns.

An object of class \code{data.frame} with 30865 rows and 627 columns.

An object of class \code{data.frame} with 30865 rows and 627 columns.

An object of class \code{data.frame} with 30865 rows and 627 columns.

An object of class \code{data.frame} with 30865 rows and 627 columns.

An object of class \code{data.frame} with 30865 rows and 627 columns.

An object of class \code{data.frame} with 30865 rows and 627 columns.

An object of class \code{data.frame} with 30865 rows and 627 columns.

An object of class \code{data.frame} with 30865 rows and 627 columns.

An object of class \code{data.frame} with 30865 rows and 627 columns.
}
\usage{
featureCount_sub1

featureCount_sub1

featureCount_sub1

featureCount_sub1

featureCount_sub1

featureCount_sub1

featureCount_sub1

featureCount_sub1

featureCount_sub1

featureCount_sub1

featureCount_sub1

featureCount_sub1

featureCount_sub1

featureCount_sub1

featureCount_sub1

featureCount_sub1

featureCount_sub1

featureCount_sub1

featureCount_sub1
}
\description{
featureCount_sub1 - a matrix of feature counts for random subset  of samples balanced according to pam50 category
rows are feature; columns are samples.

featureCount_sub1 - a matrix of feature counts for random subset  of samples balanced according to pam50 category
rows are feature; columns are samples.

featureCount_sub1 - a matrix of feature counts for random subset  of samples balanced according to pam50 category
rows are feature; columns are samples.

featureCount_sub1 - a matrix of feature counts for random subset  of samples balanced according to pam50 category
rows are feature; columns are samples.

featureCount_sub1 - a matrix of feature counts for random subset  of samples balanced according to pam50 category
rows are feature; columns are samples.

featureCount_sub1 - a matrix of feature counts for random subset  of samples balanced according to pam50 category
rows are feature; columns are samples.

featureCount_sub1 - a matrix of feature counts for random subset  of samples balanced according to pam50 category
rows are feature; columns are samples.

featureCount_sub1 - a matrix of feature counts for random subset  of samples balanced according to pam50 category
rows are feature; columns are samples.

featureCount_sub1 - a matrix of feature counts for random subset  of samples balanced according to pam50 category
rows are feature; columns are samples.

featureCount_sub1 - a matrix of feature counts for random subset  of samples balanced according to pam50 category
rows are feature; columns are samples.

featureCount_sub1 - a matrix of feature counts for random subset  of samples balanced according to pam50 category
rows are feature; columns are samples.

featureCount_sub1 - a matrix of feature counts for random subset  of samples balanced according to pam50 category
rows are feature; columns are samples.

featureCount_sub1 - a matrix of feature counts for random subset  of samples balanced according to pam50 category
rows are feature; columns are samples.

featureCount_sub1 - a matrix of feature counts for random subset  of samples balanced according to pam50 category
rows are feature; columns are samples.

featureCount_sub1 - a matrix of feature counts for random subset  of samples balanced according to pam50 category
rows are feature; columns are samples.

featureCount_sub1 - a matrix of feature counts for random subset  of samples balanced according to pam50 category
rows are feature; columns are samples.

featureCount_sub1 - a matrix of feature counts for random subset  of samples balanced according to pam50 category
rows are feature; columns are samples.

featureCount_sub1 - a matrix of feature counts for random subset  of samples balanced according to pam50 category
rows are feature; columns are samples.

featureCount_sub1 - a matrix of feature counts for random subset  of samples balanced according to pam50 category
rows are feature; columns are samples.
}
\keyword{datasets}
