% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneExpression_repl2}
\alias{geneExpression_repl2}
\title{A matrix of gene expression for replicated samples - replicate 2}
\format{
An object of class \code{data.frame} with 30865 rows and 136 columns.

An object of class \code{data.frame} with 30865 rows and 136 columns.
}
\usage{
geneExpression_repl2

geneExpression_repl2
}
\description{
geneExpression_repl2 - a matrix of gene expression for replicated samples - replicate 1
rows are feature; columns are samples.

geneExpression_repl2 - a matrix of gene expression for replicated samples - replicate 1
rows are feature; columns are samples.
}
\keyword{datasets}
