% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneExpression_sub4}
\alias{geneExpression_sub4}
\title{A matrix of gene expression for a random subset  of samples balanced according to pam50 category}
\format{
An object of class \code{data.frame} with 30865 rows and 629 columns.
}
\usage{
geneExpression_sub4
}
\description{
geneExpression_sub4 - a matrix of gene expression for random subset  of samples balanced according to pam50 category
rows are feature; columns are samples.
}
\keyword{datasets}
