% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes_annot}
\alias{genes_annot}
\title{A data frame describing the features}
\format{
genes_annot is a data frame with 30865 rows and 8 variables:
\describe{
item{geneSymbol}{             : chr  "5_8S_rRNA" "5S_rRNA" "6M1-18" "7M1-2" ...}
item{strand}{                 : Factor w/ 3 levels "-","+",".": NA NA NA NA NA 1 NA 1 1 NA ...}
item{gene_id}{                : chr  NA NA NA NA ...}
item{knownToRefSeq}{          : chr  NA NA NA NA ...}
item{knownIsoformsClusterID}{ : int  NA NA NA NA NA 22715 NA 4592 8667 NA ...}
item{chr}{                    : chr  NA NA NA NA ...}
item{start}{                  : int  NA NA NA NA NA 58346805 NA 50799408 9067707 NA ...}
item{end}{                    : int  NA NA NA NA NA 58353499 NA 50885675 9115962 NA ...}
}
}
\usage{
genes_annot
}
\description{
Data frame with 30865 rows and 8 variables describing features
}
\keyword{datasets}
