% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE96058_data_processing}
\alias{GSE96058_data_processing}
\title{A data frame with one column describing data processing in GSE96058}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 10 rows and 1 columns.
}
\usage{
GSE96058_data_processing
}
\description{
A data frame with one column describing data processing in GSE96058
}
\keyword{datasets}
