% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE96058_geneExpression_repl1}
\alias{GSE96058_geneExpression_repl1}
\title{A matrix of gene expression for replicated samples - replicate 1}
\format{
An object of class \code{data.frame} with 30865 rows and 136 columns.
}
\usage{
GSE96058_geneExpression_repl1
}
\description{
GSE96058_geneExpression_repl1 - a matrix of gene expression for replicated samples - replicate 1
rows are feature; columns are samples.
}
\keyword{datasets}
