% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE96058_geneExpression_sub2}
\alias{GSE96058_geneExpression_sub2}
\title{A matrix of gene expression for a random subset  of samples balanced according to pam50 category}
\format{
An object of class \code{data.frame} with 30865 rows and 627 columns.
}
\usage{
GSE96058_geneExpression_sub2
}
\description{
GSE96058_geneExpression_sub2 - a matrix of gene expression for random subset  of samples balanced according to pam50 category
rows are feature; columns are samples.
}
\keyword{datasets}
